requirements.txt
regex>=2024.5.15
llama-cpp-python>=0.2.84
